Moving general bounding box operations to a common module
