+++ presidio-image-redactor/presidio_image_redactor/image_redactor_engine.py
@@ -3,6 +3,7 @@ from typing import Union, Tuple, Optional
 from PIL import Image, ImageDraw, ImageChops
 
 from presidio_image_redactor import ImageAnalyzerEngine
+from presidio_image_redactor.bbox import BboxProcessor
 
 
 class ImageRedactorEngine:
@@ -16,6 +17,8 @@ class ImageRedactorEngine:
             self.image_analyzer_engine = ImageAnalyzerEngine()
         else:
             self.image_analyzer_engine = image_analyzer_engine
+
+        self.bbox = BboxProcessor()
 
     def redact(
         self,
