+++ presidio-image-redactor/tests/test_dicom_image_redactor_engine.py
@@ -5,10 +5,7 @@ import os
 import numpy as np
 from PIL import Image
 import pydicom
-from presidio_image_redactor.dicom_image_redactor_engine import DicomImageRedactorEngine
-from presidio_image_redactor.entities.image_recognizer_result import (
-    ImageRecognizerResult,
-)
+from presidio_image_redactor import DicomImageRedactorEngine, BboxProcessor
 from typing import Union, Tuple
 import pytest
 
@@ -829,185 +826,6 @@ def test_make_phi_list_happy_path(
 
 
 # ------------------------------------------------------
-# DicomImageRedactorEngine._get_bboxes_from_analyzer_results()
-# ------------------------------------------------------
-@pytest.mark.parametrize(
-    "analyzer_results, expected_bboxes",
-    [
-        (
-            [
-                ImageRecognizerResult(
-                    entity_type="TYPE_1",
-                    start=0,
-                    end=0,
-                    left=25,
-                    top=25,
-                    width=100,
-                    height=100,
-                    score=0.99,
-                ),
-                ImageRecognizerResult(
-                    entity_type="TYPE_2",
-                    start=10,
-                    end=10,
-                    left=25,
-                    top=49,
-                    width=75,
-                    height=51,
-                    score=0.7,
-                ),
-                ImageRecognizerResult(
-                    entity_type="TYPE_3",
-                    start=25,
-                    end=35,
-                    left=613,
-                    top=26,
-                    width=226,
-                    height=35,
-                    score=0.6,
-                ),
-            ],
-            [
-                {
-                    "entity_type": "TYPE_1",
-                    "score": 0.99,
-                    "left": 25,
-                    "top": 25,
-                    "width": 100,
-                    "height": 100,
-                },
-                {
-                    "entity_type": "TYPE_2",
-                    "score": 0.7,
-                    "left": 25,
-                    "top": 49,
-                    "width": 75,
-                    "height": 51,
-                },
-                {
-                    "entity_type": "TYPE_3",
-                    "score": 0.6,
-                    "left": 613,
-                    "top": 26,
-                    "width": 226,
-                    "height": 35,
-                },
-            ],
-        ),
-    ],
-)
-def test_get_bboxes_from_analyzer_results_happy_path(
-    mock_engine: DicomImageRedactorEngine,
-    analyzer_results: list,
-    expected_bboxes: list,
-):
-    """Test happy path for DicomImageRedactorEngine._get_bboxes_from_analyzer_results
-
-    Args:
-        analyzer_results (list): Results from using ImageAnalyzerEngine.
-        expected_bboxes (list): Expected output bounding box list.
-    """
-    # Arrange
-
-    # Act
-    test_bboxes = mock_engine._get_bboxes_from_analyzer_results(analyzer_results)
-
-    # Assert
-    assert test_bboxes == expected_bboxes
-
-
-# ------------------------------------------------------
-# DicomImageRedactorEngine._format_bboxes()
-# ------------------------------------------------------
-@pytest.mark.parametrize(
-    "mock_intermediate_bbox, padding_width, expected_bboxes",
-    [
-        (
-            [
-                {
-                    "entity_type": "TYPE_1",
-                    "score": 0.99,
-                    "left": 10,
-                    "top": 15,
-                    "width": 100,
-                    "height": 100,
-                },
-                {
-                    "entity_type": "TYPE_2",
-                    "score": 0.7,
-                    "left": 25,
-                    "top": 49,
-                    "width": 75,
-                    "height": 51,
-                },
-                {
-                    "entity_type": "TYPE_3",
-                    "score": 0.6,
-                    "left": 613,
-                    "top": 26,
-                    "width": 226,
-                    "height": 35,
-                },
-            ],
-            25,
-            [
-                {"top": 0, "left": 0, "width": 100, "height": 100},
-                {"top": 24, "left": 0, "width": 75, "height": 51},
-                {"top": 1, "left": 588, "width": 226, "height": 35},
-            ],
-        ),
-    ],
-)
-def test_format_bboxes_happy_path(
-    mocker,
-    mock_engine: DicomImageRedactorEngine,
-    mock_intermediate_bbox: dict,
-    padding_width: int,
-    expected_bboxes: list,
-):
-    """Test happy path for DicomImageRedactorEngine._format_bboxes
-
-    Args:
-        mock_intermediate_bbox (dict): Value for mock of _get_bboxes_from_analyzer_results.
-        padding_width (int): Pixel width used for padding.
-        expected_bboxes_dict (dict): Expected output bounding box dictionary.
-    """
-    # Arrange
-    mock_get_bboxes = mocker.patch(
-        "presidio_image_redactor.dicom_image_redactor_engine.DicomImageRedactorEngine._get_bboxes_from_analyzer_results",
-        return_value=mock_intermediate_bbox,
-    )
-
-    # Act
-    test_bboxes_dict = mock_engine._format_bboxes([], padding_width)
-
-    # Assert
-    assert mock_get_bboxes.call_count == 1
-    assert test_bboxes_dict == expected_bboxes
-
-
-@pytest.mark.parametrize(
-    "padding_width, expected_error_type",
-    [(-1, "ValueError"), (-200, "ValueError")],
-)
-def test_format_bboxes_exceptions(
-    mock_engine: DicomImageRedactorEngine, padding_width: int, expected_error_type: str
-):
-    """Test error handling of _format_bboxes
-
-    Args:
-        padding_width (int): Pixel width used for padding.
-        expected_error_type (str): Type of error we expect to be raised.
-    """
-    with pytest.raises(Exception) as exc_info:
-        # Act
-        _ = mock_engine._format_bboxes([], padding_width)
-
-    # Assert
-    assert expected_error_type == exc_info.typename
-
-
-# ------------------------------------------------------
 # DicomImageRedactorEngine._set_bbox_color()
 # ------------------------------------------------------
 @pytest.mark.parametrize(
@@ -1259,8 +1077,8 @@ def test_DicomImageRedactorEngine_redact_happy_path(
     )
 
     mock_format_bboxes = mocker.patch.object(
-        DicomImageRedactorEngine,
-        "_format_bboxes",
+        BboxProcessor,
+        "format_bboxes_for_dicom",
         return_value=None,
     )
 
@@ -1356,7 +1174,7 @@ def test_DicomImageRedactorEngine_redact_single_dicom_image_happy_path(
     )
 
     mock_format_bboxes = mocker.patch(
-        "presidio_image_redactor.dicom_image_redactor_engine.DicomImageRedactorEngine._format_bboxes",
+        "presidio_image_redactor.BboxProcessor.format_bboxes_for_dicom",
         return_value=None,
     )
